requirements.txt
aiofiles>=23.2.1
aiogram==3.6.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
redis>=5.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
PyPDF2>=3.0.0
pandas>=2.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
reportlab>=4.0.0
icalendar>=5.0.0
anthropic>=0.20.0
pytz>=2023.3
openpyxl>=3.1.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
matplotlib>=3.7.0
python-dateutil>=2.8.2
